# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Odace Data Pipeline API",
    "description": "\n## Odace Data Pipeline Platform\n\nA cloud-based data pipeline platform for managing data ingestion, transformation, and processing.\n\n### Features\n\n* **Multi-layer architecture**: Bronze (ingestion), Silver (transformation), Gold (aggregation)\n* **Delta Lake storage**: Efficient data lake with ACID transactions\n* **API-driven**: RESTful API for pipeline orchestration\n* **Authenticated**: Secure API key authentication\n* **Scalable**: Runs on Google Cloud Run\n\n### Authentication\n\nAll endpoints require an API key in the Authorization header:\n\n```\nAuthorization: Bearer sk_live_YOUR_API_KEY\n```\n\n### Available Pipelines\n\n**Bronze Layer (Data Ingestion)**:\n- accueillants, geo, logement, gares, lignes, zones_attraction\n\n**Silver Layer (Data Transformation)**:\n- accueillants, geo, logement, gares, lignes, zones_attraction\n\n### Getting Started\n\n1. Obtain an API key from your administrator\n2. Include the API key in all requests: `Authorization: Bearer sk_live_...`\n3. Use `/api/pipeline/list` to see available pipelines\n4. Run pipelines with `/api/bronze/{pipeline_name}` or `/api/silver/{pipeline_name}`\n\nFor detailed documentation, visit the project repository.\n    ",
    "contact": {
      "name": "Odace Data Team",
      "email": "data@odace.example.com"
    },
    "license": {
      "name": "Internal Use Only"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/admin/api-keys": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List All Api Keys",
        "description": "List all API keys (without plaintext keys).\n\nRequires admin authentication via ADMIN_SECRET.\n\nArgs:\n    db: Firestore async client\n    \nReturns:\n    List of API key metadata",
        "operationId": "list_all_api_keys_admin_api_keys_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/APIKeyInfo"
                  },
                  "type": "array",
                  "title": "Response List All Api Keys Admin Api Keys Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create New Api Key",
        "description": "Create a new API key for a user.\n\nRequires admin authentication via ADMIN_SECRET.\n\nArgs:\n    request: Contains user_id for the new API key\n    db: Firestore async client\n    \nReturns:\n    The newly created API key (shown only once)",
        "operationId": "create_new_api_key_admin_api_keys_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAPIKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAPIKeyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/admin/api-keys/revoke": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Revoke Existing Api Key",
        "description": "Revoke an API key (soft delete - sets active=False).\n\nRequires admin authentication via ADMIN_SECRET.\n\nArgs:\n    request: Contains the API key to revoke\n    db: Firestore async client\n    \nReturns:\n    Success message",
        "operationId": "revoke_existing_api_key_admin_api_keys_revoke_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeAPIKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/admin/api-keys/delete": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Delete Existing Api Key",
        "description": "Permanently delete an API key from Firestore.\n\nRequires admin authentication via ADMIN_SECRET.\n\nArgs:\n    request: Contains the API key to delete\n    db: Firestore async client\n    \nReturns:\n    Success message",
        "operationId": "delete_existing_api_key_admin_api_keys_delete_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAPIKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIKeyInfo": {
        "properties": {
          "hash": {
            "type": "string",
            "title": "Hash"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          }
        },
        "type": "object",
        "required": [
          "hash",
          "user_id",
          "created_at",
          "last_used_at",
          "active"
        ],
        "title": "APIKeyInfo",
        "description": "API key metadata (without plaintext key)."
      },
...
      "CreateAPIKeyRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "CreateAPIKeyRequest",
        "description": "Request model for creating a new API key."
      },
      "CreateAPIKeyResponse": {
        "properties": {
          "api_key": {
            "type": "string",
            "title": "Api Key"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "API key created successfully. Save this key - it will not be shown again."
          }
        },
        "type": "object",
        "required": [
          "api_key",
          "user_id",
          "created_at"
        ],
        "title": "CreateAPIKeyResponse",
        "description": "Response model for API key creation."
      },
      "DeleteAPIKeyRequest": {
        "properties": {
          "api_key": {
            "type": "string",
            "title": "Api Key"
          }
        },
        "type": "object",
        "required": [
          "api_key"
        ],
        "title": "DeleteAPIKeyRequest",
        "description": "Request model for deleting an API key."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RevokeAPIKeyRequest": {
        "properties": {
          "api_key": {
            "type": "string",
            "title": "Api Key"
          }
        },
        "type": "object",
        "required": [
          "api_key"
        ],
        "title": "RevokeAPIKeyRequest",
        "description": "Request model for revoking an API key."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}